+++ plugins/readme-backend/package.json
@@ -24,19 +24,23 @@
   },
   "dependencies": {
     "@backstage/backend-common": "^0.19.8",
+    "@backstage/catalog-client": "^1.4.6",
+    "@backstage/catalog-model": "^1.4.3",
     "@backstage/config": "^1.1.1",
+    "@backstage/integration": "^1.7.2",
+    "@backstage/plugin-permission-common": "^0.7.10",
     "@types/express": "*",
     "express": "^4.17.1",
     "express-promise-router": "^4.1.0",
-    "winston": "^3.2.1",
     "node-fetch": "^2.6.7",
+    "winston": "^3.2.1",
     "yn": "^4.0.0"
   },
   "devDependencies": {
     "@backstage/cli": "workspace:^",
     "@types/supertest": "^2.0.12",
-    "supertest": "^6.2.4",
-    "msw": "^1.0.0"
+    "msw": "^1.0.0",
+    "supertest": "^6.2.4"
   },
   "files": [
     "dist"
